Update README.md
This began as a programming assessment task for my Masters degree. I re-worked it in the interest of improving my coding skills and included a little bit of statistics.

It looks at ratings in the context of how far the author is into their career (by years and number of books published), the price and length of the book, the genre, and whether it has won an award. There is then a quick look at the ratings of books compared to their adapted movies.

There are a couple of nice and surprising results!